Fixes xprocessing jar to properly handle shading.
The issue was due to incorrect handling of shading in jarjar for kotlin metadata. In particular, jarjar does not replace UTF8 usages in the classfile, so the Strings in the kotlin metadata do not get replaced. This breaks when the class is compiled with kotlinc which uses the kotlin metadata during compilation.

To fix this, I've switched to using shadowJar to do the shading (https://imperceptiblethoughts.com/shadow/).

Closes #2765

RELNOTES=Fix #2765: Fixes xprocessing jar to properly handle shading.
PiperOrigin-RevId: 386099651