Fix link in Dagger testing.
In addition, I've changed the title of the linked section. In particular, the link takes the user to a section called "Downsides of the alternative", but it's not clear what "alternative" means here with out the context of the previous section. Thus, I changed the title to "Downsides of manual instantiation" to make it more explicit.

RELNOTES=N/A
PiperOrigin-RevId: 414869149