Collect Hilt aggregating task classpath via separate configurations.
Instead of using a single Hilt created configuration to apply the aggregated transform, only use the Hilt configuration to transform the JavaCompile's classpath and output and combine it with a resolved runtime configuration that also applies the transform. This leads to Hilt no longer having to use AGP attributes to configure the configuration making it more decoupled and also avoids the issue of trying to jetify twice.

Moreover change the request artifacts from processed-jar to android-classes since Hilt is only interested in classes and not resources.

RELNOTES=n/a
PiperOrigin-RevId: 438361718