Fixed the incorrect logic from ProductionBindingRepresentation which could have resulted in type casting error.
In the existing logic, if someone is requesting a production binding with Provider, Dagger will end up generating a Producer typed expression.

This cl changes the behavior, so that if a binding request kind for a production binding is PROVIDER, then use derivedFromFrameworkInstanceRequestRepresentation to derive the expression from the corresponding ProducerNodeInstanceRequestRepresentation. This will still fail, but with better error message "request kind PROVIDER cannot be satisfied by production binding" from the BindingGraphPlugin Validator.

RELNOTES=n/a
PiperOrigin-RevId: 629639958