Update Bazel to use JDK 18.
This is required before updating to the latest XProcessing jars (which require JDK 17+). Note that we're not using JDK 17 because there's a bug that affects Dagger (https://bugs.openjdk.org/browse/JDK-8268575).

This CL sets the default target level to 8, which is used by all of the processors, but some of the Dagger (non-processor) artifacts are set explicitly to target 7. Note that we're keeping the target low to avoid forcing downstream users to have to use a higher JDK.

I've also added a validation step to our GitHub Actions to add some level of guarantee that the classes in the artifact jars all have the expected language level.

RELNOTES=N/A
PiperOrigin-RevId: 658131076