[Cleanup]: Use `MultiboundMapBinding` and `MultiboundSetBinding` in more places.
This CL replaces some usages of `Binding`/`ContributionBinding` with the more explicit types when possible. This has the benefit of being able to use more specific APIs in these locations and also avoids having to check that the bindings are of the correct kind in every method.

RELNOTES=N/A
PiperOrigin-RevId: 663836816