Remove support for Java 7.
We're aiming to add proper nullable annotations to Dagger's runtime API types in the upcoming Dagger release. In order to do this we need to drop support for Java 7.

Given Java 7 itself was officially out of support after July 2022, dropping support for Java 7 in Dagger seem reasonable.

RELNOTES=Remove support for Java 7
PiperOrigin-RevId: 668525969