Ban scoping on `@Binds` that delegate to `@Produces` implementations.
This CL bans scoping on `@Binds` methods that delegate to `@Produces` methods. Note that scoping on `@Produces` methods themselves is already banned because production bindings are implicitly scoped.

There's a few benefits to banning this. First, allowing users to arbitrarily add scopes can be misleading because the scopes will just be ignored. Second, this change allows us to clean up places in `BindingFactory` that required knowledge of the `BindingType` (which is something we're currently trying to remove).

RELNOTES=Ban scoping on production bindings.
PiperOrigin-RevId: 673020697