Refactor `BindingGraphFactory` to build network directly.
This CL removes the need for `BindingGraphConverter` and `LegacyBindingGraph` and builds up the network as part of the resolution. Since the `BindingGraphConverter` is now only used by `LegacyBindingGraphFactory` I've renamed it to `LegacyBindingGraphConverter`.

Note: This has been a change we've wanted to make for a while now but hasn't been possible due to the complexity and brittleness of the resolution logic.
RELNOTES=N/A
PiperOrigin-RevId: 675217954