[Refactor]: Split BindingFactory.syntheticOptionalDeclaration into separate methods for present vs absent.
This change is similar to other refactors which split the `multibinding` method into `multiboundSet` and `multiboundMap`. In general, I think it's better if `BindingFactory` is just responsible for creating the bindings and does not have logic to decide on which binding it should create. This also makes it clearer at the call site which binding we're actually using.

RELNOTES=N/A
PiperOrigin-RevId: 676053984