[Refactor]: Fork ResolvedBindings for LegacyBindingGraphFactory.
ResolvedBindings is currently shared by both BindingGraphFactory and LegacyBindingGraphFactory, but future changes to BindingGraphFactory will need to make changes to ResolvedBindings that don't apply to  LegacyBindingGraphFactory so I'm forking this in advance to make the legacy code independent of future changes.

I've also removed the `LegacyBindingGraphConverter.LegacyBindingGraph` interface. This was previously needed since both `BindingGraphFactory` and `LegacyBindingGraphFactory` had their own implementations of this interface, but now that `BindingGraphFactory` no longer uses it the interface is no longer needed.

RELNOTES=N/A
PiperOrigin-RevId: 676123197