Refactor BindingGraphFactory.RequiresResolutionChecker to use the network directly.
This is a follow-up to CL/675217954, which changed `BindingGraphFactory` to build the network during the resolution stage. Now that BindingGraphFactory builds up a network, we can use that network directly in `RequiresResolutionChecker` when calculating whether a particular `BindingNode` needs to be re-resolved in the current component.

This also fixes a bug (b/367426609) that was hiding in the old logic and is now fixed by iterating via the network directly.

RELNOTES=N/A
PiperOrigin-RevId: 676530675