Move unwrapping of delegate keys to a single location.
This CL moves the unwrapping of the key directly into `ComponentDeclarations#delegates(Key)` instead of requiring all callers to remember to unwrap the key individually. This change will make this API less error-prone, as it's already caused one bug (b/147020838).

Note: In order to make this change, I needed to refactor how we handled the strictMultibindingValidation flag. Rather than having the flag control whether or not we unwrap the key, the flag now just exits early from `resolverContainsDelegateDeclarationForBinding`, which should be functionally equivalent to how it works today. In particular, the delegate map multibindings are always wrapped today, so not unwrapping the key will always result in a miss when looking up the key in the map.
RELNOTES=N/A
PiperOrigin-RevId: 681473673