Remove the `ignorePrivateAndStaticInjectionForComponent` compiler option.
This option was only used by the TCK tests, so I've replaced the compiler option with a hard-coded check for the TCK package. As was stated in the javadoc for the flag, it was not meant to be used for other purposes, so removing it should be fine.

This change also fixes a some memory bugs in how `InjectValidator` was previously implemented.

  1. Now, only a single validator instance is used in the common case where
     `privateMemberDiagnostic`/`staticMemberDiagnostic` have the default
     value (`ERROR`). Previously, this resulted in at least 3 instances being
     used, which duplicated caching across all three instances
  2. Now, both `validator` and `validatorWhenGeneratingCode` have their caches
     cleared between rounds. Previously, only `validator` had its cache cleared
     because `validatorWhenGeneratingCode` was being created and managed
     outside Dagger. This also created a memory leak in one case because it was
     used within a scoped binding and it's cache was never cleared.

RELNOTES=Remove the `ignorePrivateAndStaticInjectionForComponent` compiler option.
PiperOrigin-RevId: 686174065