[Refactor]: Change `FrameworkField#create()` to better align with the XPoet migration.
During the XPoet migrations `FrameworkField#create()` will be called by both migrated and non-migrated code.

Due to b/379934406, it's currently not feasible to generally go from JTypeName to XTypeName so I'm changing up the `create` method to take in the framework `ClassName` separately. Unlike `TypeName`, a `ClassName` can be converted to an `XClassName` which will make the interopt between the two usages easier during the migration.

This CL also updates FrameworkFieldTest to use the new API.

RELNOTES=N/A
PiperOrigin-RevId: 700078936