Initial setup of Gradle as a build system for Dagger
At the repository root, the root project 'dagger-parent' configures common build functionality and includes the various projects that will be placed 'gradle-projects', for now only the runtime is configured.

The buildSrc directory is a special Gradle sub-project to share common build logic between sub-projects, it will become more useful in time as the projects gets more complex and shared build logic is needed.

Dependencies and versions are defined in a Version Catalog in gradle/libs.versions.toml, these will eventually need to be validated to be in-sync with Bazel's WORKSPACE.

RELNOTES=N/A
PiperOrigin-RevId: 710702061