Add binary compatibility validator to the Gradle projects
The validator (https://github.com/Kotlin/binary-compatibility-validator) will help us track public API changes and validate that the conversion from Java to Kotlin does not cause any breakage.

API validation is part of the Gradle build process and will fail the build if API changes are done but not reflected in the API file. To update the API file execute `./gradlew apiDump` and commit the .api file for review along with the code changes.

RELNOTES=N/A
PiperOrigin-RevId: 713662828