Don't store XElement outside of processing rounds.
`GeneratesRootInputProcessorTest` stores the `List<XElement>` returned by `GeneratesRootInputs.getElementsToWaitFor()` to test the output. However, this causes an issue with the latest XProcessing drop when trying to access the fully qualified name of the element outside of the actual processing rounds. This is fixed if we replaced the `List<XElement>` with `List<String>` and calculate the qualified name immediately inside the processing round.

RELNOTES=N/A
PiperOrigin-RevId: 720687140