Remove @AutoValue for cases where we're not actually relying on equals/hashcode.
This CL removes AutoValue from cases where we shouldn't be relying on equals/hashcode of these types.

This is part of a cleanup related to the XPoet migration, in an attempt to be more intentional about what we're relying on in our AutoValue classes. In particular, cases where we're using `TypeName` as part of the equals/hashcode for AutoValue will likely be incorrect/error-prone if switched over to `XTypeName` since the latter takes into account nullability and mutability. This CL narrows these cases to at least only includes cases where we actually require equals/hashcode.

RELNOTES=N/A
PiperOrigin-RevId: 725789612