Migrate remaining TypeNames to XTypeNames.
This CL migrates all `TypeNames` usages to `XTypeNames` and removes the old util class. I've also converted XTypeNames to Kotlin because there are some APIs that are not yet annotated with `@JvmField`/`@JvmStatic` so they are not accessible to Java sources. Converting to Kotlin allows us to access those APIs without having to go through the tedious process of updating XPoet and then Dagger's XPoet jars.

RELNOTES=N/A
PiperOrigin-RevId: 731341921