Remove the bazel build command from bazel-build action
Due to the restructuring of Dagger's sources, running `bazel build //...` will build both java and javatests sources, and this ends up taking much more time (7min -> 27min) and memory (0.4GB -> 2.1GB) in GitHub Actions than before. Since this task is a bottleneck to many other tasks, e.g. the various gradle tests, I've removed the build command since everything we need to build the artifacts should be built when generating the artifacts.

Furthermore, the `bazel test //...` in bazel-test action should already cover everything that was being built in the bazel-build action.

As a follow-up we should probably rename the "bazel-build" action to something like "build-artifacts" to account for this change in behavior.

RELNOTES=N/A
PiperOrigin-RevId: 731479252