Fix crash in KSP2 when using nullable return type in AssistedFactory create method.
Dagger will crash when a nullable type is passed to `constructor.asMemberOf(assistedInjectType)`. This CL uses `assistedInjectType.makeNonNullable()` to fix this issue and allow users to return nullable type in KSP2. We decided to allow this because returning a nullable type is always technically okay even if the return type is never nullable, and we're allowing flexibility for cases where the interface may be shared with cases other than assisted factory that may return null.

See b/395002429 for more details.

RELNOTES=N/A
PiperOrigin-RevId: 734271287