[XPoet Migration]: Migrate CodeBlock usages to XCodeBlock.
Note: XPoet adds a parenthesis around all of the casted codeblocks (since there's not a general way to determine if it's needed). Since Dagger keeps track of whether the parenthesis are needed, we could add an input parameter to control this in `XCodeBlock#ofCast()`; however, there doesn't seem to be much harm in adding the parenthesis anyway.
RELNOTES=N/A
PiperOrigin-RevId: 739976457