Remove #delayErrors() from processing step.
This feature was originally added to allow Hilt to continue processing and potentially generate files that would otherwise show up as missing symbols and potentially pollute the error logs. However, the implementation of this feature doesn't work with KSP2 which will crash if elements are accessed outside of the round they were created. We could work around this issue by storing identifying information about elements (e.g. parameter name -> method signature -> class name) to recreate elements in the current round; but we decided to just remove this feature since it's not clear it has a lot of benefit.

RELNOTES=N/A
PiperOrigin-RevId: 741279228