[XPoet Migration]: Migrate FieldSpec to XPropertySpec.
This CL adds a new utility class, XPropertySpecs.java, which includes a wrapper of XPropertySpec.Builder to expose an API that's easier for migration. In particular, the wrapper is very similar to JavaPoet's API, and it allows setting input as both JavaPoet and/or XPoet.

The JavaPoet APIs are marked as @Deprecated and they will be cleaned up in a future CL.

RELNOTES=N/A
PiperOrigin-RevId: 748332454