[XPoet Migration]: Replace TypeName usages in AutoValue with Equivalence.Wrapper<XType>.
As part of the XPoet migration, I've been replacing `TypeName` with `XTypeName`. However, that doesn't apply to cases where we currently rely on the equals/hashcode of `TypeName`, e.g. AutoValue classes. In these cases, I'm replacing `TypeName` with an `Equivalence.Wrapper<XType>` that has the same equals/hashcode as `TypeName`. This change should be a noop.

RELNOTES=N/A
PiperOrigin-RevId: 750346029