[XPoet Migration]: Remove workaround to control @Overrides ordering.
This CL removes the workaround that lets us control the ordering of the `@Overrides` annotation and updates the golden files. This workaround was initially used just to avoid causing changes to our golden files while the migration was in progress, but it's being removed now that the migration is nearing completion.

RELNOTES=N/A
PiperOrigin-RevId: 752395970