Add dagger-android-processor to Gradle projects
This required introducing a new configuration called 'unshaded' so that in this Gradle multi-project setup the unshaded version of Dagger artifacts can be used as as dependencies. This allows dagger-android-processor to depend on dagger-compiler and its XProcessing utilities using the non-shaded signatures. Upon publishing dagger-android-processor goes through shading, relocating symbol references, so that it matches the shaded xprocessing in dagger-spi and the also relocated signatures in dagger-compiler.

Also make output jars reproducible as suggested in https://gradleup.com/shadow/configuration/reproducible-builds/

RELNOTES=N/A
PiperOrigin-RevId: 761054160