Fix ActivityGenerator's logic for onCreate modifiers..
This CL fixes an issue with ActivityGenerator which was looking for modifiers on the `@AndroidEntryPoint`-annotated class's `onCreate` method rather than the super class method. See b/418908653 for details.

I've also had to update a few places where we were missing the `@Nullable` annotation due to the same bug.

RELNOTES=N/A
PiperOrigin-RevId: 761597753