Support using `java.util.Set`/`java.util.Map` for multibindings in kotlin sources.
In general, users should be using `kotlin.collections.Set` rather than `java.util.Set` in kotlin sources, but we've decided we should still support it. This also keeps things consistent with KAPT, since we support it implicitly in KAPT today.

This CL also reenables KSP in Dagger's functional tests, which was accidentally turned off when switching `dagger.useKsp=true` from plugin options to javacopts since we were looking for it in plugin options.

RELNOTES=N/A
PiperOrigin-RevId: 761741915