Fix duplicate xprocessing shaded deps.
This CL removes xprocessing from the dagger-spi artifact and moves it into dagger-compiler.

For further safety, I've limited the visibility of all dagger-compiler libraries and added alias targets under `/dagger-compiler:*` to expose libraries that are used outside of the artifact. These aliases add the artifact id to the tag to ensure any dependent properly detects the dagger-compiler dependency.

RELNOTES=N/A
PiperOrigin-RevId: 762442065