Separate compiler tests into a separate macro.
This CL separates compiler tests into a separate macro from functional tests. The main reason for this is that the macro for compiler tests are much simpler than functional tests because we don't need to run then with multiple variants (e.g. with/without KSP or various build flags) since those things are set manually for each test case in a compiler test.

This also simplifies the functional test macros because we can remove the `functional` parameters and checks in these macros.

RELNOTES=N/A
PiperOrigin-RevId: 768208400