Move `Accessibility.java` into `/xprocessing`.
The `/langmodel` package was initially used for extensions of `javax.lang.model`. However, since we've migrated to XProcessing `Accessibility.java` is the only remaining class in this package, and doesn't really make sense anymore.

I've also migrated `AccessibilityTest.java` to XProcessing and moved it out of the `langmodel` package. (Previously, this test was actually being skipped because there was no build file in the directory).

RELNOTES=N/A
PiperOrigin-RevId: 769773879