Fix issue with inaccessible type referenced in generated factory.
This CL updates the logic in `SourceFiles#generateBindingFieldsForDependencies()` to account for the visibility of the dependency type. Previously, we were assuming all dependencies were accessible, which isn't true in cases where the factory may need a dependency due to a super type's member injected field/method whose type isn't public.

RELNOTES=N/A
PiperOrigin-RevId: 773005998