Avoid javac stack overflow error when building AST of generated factory.
This CL moves the creation of the `MapFactory` and `SetFactory` into a separate initialization method and also separates each `put`/`add` call as a separate statement. Having each `put`/`add` call as a separate statement helps reduce the size of the stack when javac computes the AST for this code.

RELNOTES=N/A
PiperOrigin-RevId: 781107611