Replace `ProducerEntryPointView` with a `FrameworkFieldInitializer`.
The `ProducerEntryPointView` creates a framework field for a producer entry point but doesn't go through the standard `FrameworkFieldInitializer` to do so. This CL refactors the code to use a `FrameworkFieldInitializer`.

The main benefit of this CL is that it ensures all of the framework fields are initialized in the same way rather than bespoke implementation scattered throughout the codebase. This will be particularly useful when we start generating kotlin since that will requires changes to the framework fields (e.g. they can't use raw types) and with this refactor we won't have to make that change in multiple locations.

RELNOTES=N/A
PiperOrigin-RevId: 781197268