[Kotlin codegen]: Put `static` class members in a companion object for Kotlin codegen.
For Kotlin codegen there is no "static" modifier.

This CL modifies `XTypeSpecs.Builder` to keep track of members marked as "static" and then adds them to a companion object during `build()`.

Note: this change affects some of our java codegen golden files because it reorganizes all static methods to be at the bottom of the class.
RELNOTES=N/A
PiperOrigin-RevId: 790796064