Remove the 'android-classes' from Dagger's artifact transform to avoid ambiguity in transform chains with Gradle 9.0.0.
The use of the attribute is not needed as AGP setups enough transforms such that an android library (project or external) has an outgoing variant whose artifact type is 'jar'.

This change also bumps the minimum required AGP version of the plugin from 8.1 to 8.4 as that is the minimum version where the removed attribute is no longer needed.

Due to the AGP update the tests projects are also updated:
* Gradle from 8.4 to 8.6
* AGP from 8.1.1 to 8.4.2
* Android build tools from 33.0.1 to 34.0.0
* Android compile and targets SDK from 33 to 34

Fixes https://github.com/google/dagger/issues/4847

RELNOTES=N/A
PiperOrigin-RevId: 800424938