Fix XMethodElements#hasOverride() bug.
This CL update the `XMethodElements#hasOverride()` to take kotlin property methods into account. The fix is to use `XConverters.toKS(XElement)` rather than `XConverters.toKS(XExecutableElement)` since the former properly handles properties while the latter will throw if the method is from a property in KSP. With this approach KAPT and KSP will return the same results.

RELNOTES=N/A
PiperOrigin-RevId: 802724734