Add a test suite to group our generated functional tests.
This CL adds a test suite to allow running all variants of a test using the it's base name.

# Background

The macros defined in `test_defs.bzl` generate multiple variants of each test that are built with different javacopts (e.g. Javac vs KSP, Java codegen vs Kotlin codegen, etc.). The name given to the macro (e.g. "BarTest") is currently used as a base name for the actual test targets which have additional suffixes appended to the name (e.g. "BarTest_Javac_JavaCodegen", "BarTest_Ksp_JavaCodegen"). However, one issue with this setup is that it is difficult to run all of the tests associated with "BarTest" since you either need to know all of the variants and run each of then individually, or use `:all` to run all of the targets in the package. This CL adds a test suite using the base name that will run all variants of that test (e.g. `blaze test foo/bar:BarTest`).

RELNOTES=N/A
PiperOrigin-RevId: 808329922