Fix issue with DaggerSuperficialValidation#validateTypeHierarchyOf().
This CL fixes an issue with `DaggerSuperficialValidation#validateTypeHierarchyOf()` that can affect `@Binds` method validation. In particular, the issue is that we don't recursively validate the type hierarchy for the type arguments. For example, if `class Foo extends Bar<Baz>` and `Baz` extends an invalid type (e.g. `class Baz extends MissingType`) then we previously didn't catch this case and we assumed the type was valid. This causes issues with `@Binds` validation because if there's an invalid type in the hierarchy the `KSType#isAssignableFrom(KSType)` check will be incorrect.

RELNOTES=Fix issue with DaggerSuperficialValidation#validateTypeHierarchyOf().
PiperOrigin-RevId: 808650035