Fix issue with incorrect variance in map multibinding method in KSP.
This issue happens for a type like:

```kotlin
@JvmSuppressWildcards Map<Key, out Value>
```

In this case, XProcessing calculates the correct variance for the complete jvm type name, `Map<Key, Value>`, however when we get the jvm type name of the value by itself we get `? extends Value` rather than `Value` because we loose the information about the `@JvmSuppressWildcards` when looking at only the value type.

The fix here is to work with the XTypeName directly, which should keep the variance information intact.

RELNOTES=N/A
PiperOrigin-RevId: 818896362